move top activity helper to ita
