+++ common/src/main/java/top/trumeet/common/ita/ITopActivity.java
@@ -1,4 +1,4 @@
-package com.xiaomi.helper;
+package top.trumeet.common.ita;
 
 import android.content.Context;
